+++ src/main/java/com/example/webprojectgames/repositories/GamesRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GamesRepository extends JpaRepository<Game,Integer > {
+public interface GamesRepository extends JpaRepository<Game,Long > {
 
 
 }
