+++ src/main/java/com/example/webprojectgames/services/GameService.java
@@ -11,7 +11,7 @@ public interface GameService {
     List<Game> getAllGames();
     List<Game> getUserGamesCollection(int userId);
 
-    Game findById(int id);
+    Game findById(long id);
 
     void saveGame(Game game);
